requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
celery==5.2.7
requests==2.28.2
redis==4.5.1
backports.zoneinfo==0.2.1
Django==4.0.0
django-admin-generator==2.2.1
django-ckeditor-5==0.2.2
django-filter==21.1
django-lifecycle==1.0.0
django-utils2==2.12.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
packaging==23.0
Pillow==9.1.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.3
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.20.0
python-utils==3.5.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urlman==2.0.1
vine==5.0.0
wrapt==1.14.0
zipp==3.14.0
django-phonenumber-field
django-phonenumbers
django-mdeditor
openai==0.27.4
pydub==0.25.1
django-ninja==0.21.0
django-abnorm==1.3.0
python-telegram-bot==20.2
git+https://github.com/tiliv/django-colors-formatter.git
git+https://github.com/Neural-Dynamics-Web/django-jazzmin.git@62dce94027a3bef74e18d7cb33bda39bcaf665ba
git+https://github.com/Neural-Dynamics-Web/django-admin-autocomplete-filter.git@a4dca9e0a260c80b1718fb3bbe09579b2523a61a